Move `Utility` to AST.
